OPTIMIZE: take marketIDMap into common dir
